修改

+++ database/migrations/2019_12_25_115034_alter_tour_alter_column_status.php
@@ -14,7 +14,7 @@ class AlterTourAlterColumnStatus extends Migration
     public function up()
     {
         Schema::table('tour', function (Blueprint $table) {
-            $table->tinyInteger('status')->default(1)->nullable()->after('warehouse_lat')->comment('状态')->change();
+            $table->smallInteger('status')->default(1)->nullable()->after('warehouse_lat')->comment('状态')->change();
         });
     }
